api/python/requirements.txt
mysql-connector-python==8.1.0
flask_cors==3.0.10
python-dotenv==0.19.0
